requirements.txt
# -------- API --------
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.6.4
pydantic-settings>=2.0.0

# -------- Rate Limiting --------
slowapi>=0.1.9

# -------- File ingestion --------
pdfplumber==0.11.4
pypdf==4.3.1
Pillow==10.4.0

# -------- Embedding engine --------
sentence-transformers==2.7.0
transformers==4.38.2
torch==2.2.2

# -------- Vector operations --------
numpy==1.26.4
faiss-cpu==1.8.0
scikit-learn==1.4.0

# -------- LLM --------
openai>=1.14.0
httpx==0.27.0
python-dotenv==1.0.1
tenacity>=8.2.0

# -------- Testing --------
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# -------- Code Quality --------
mypy>=1.4.0
ruff>=0.1.0